Modified variantgen to generate the graphics pokemon.h file.
Significant refactoring of sprite and palette lookup. Minor bugfixes
throughout. There may be other calls with have the personality and
variant in the wrong order.
The correct order is:
species
variant
personality.

+++ include/pokemon.h
@@ -438,7 +438,7 @@ void DestroyMonSpritesGfxManager(void);
 u8 *MonSpritesGfxManager_GetSpritePtr(u8 bufferId);
 
 bool8 IsTrainerRival(u16 trainerId);
-u8 GenerateMonVariant(u16 species, u32 variantSeed);
+u16 GenerateMonVariant(u16 species, u32 variantSeed);
 
 extern struct SpritePalette dynamicPalette;
 extern u16 dynamicPaletteData[16];
